add performance profile mode & channel mix mode on device

+++ packages/coast_audio_miniaudio/mabridge/src/dart_bridge.c
@@ -1,5 +1,5 @@
 #include "dart_api_dl.h"
 
 int dart_bridge_init(void* pData) {
-  return Dart_InitializeApiDL(pData);
+  return (int)Dart_InitializeApiDL(pData);
 }
